mt5_bridge/requirements.txt
# MT5 Bridge Requirements
MetaTrader5==5.0.45
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
requests==2.31.0
PyYAML==6.0.1
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
python-dateutil==2.8.2
scipy==1.11.2
scikit-learn==1.3.1
talib-binary==0.4.28
websocket-client==1.6.2
pydantic==2.4.2
